+++ RELEASES.md
@@ -6,7 +6,7 @@
 
 ### Major features
 * In mode `EF -counterexample`, IMITATOR outputs an example of parameter valuation, and a full concrete run, with a graphical representation of the clocks and discrete variables evolution. [WORK IN PROGRESS]
-* New accepting cycle synthesis (`-mode ActLoopSynth`)
+* New accepting cycle synthesis (`-mode AccLoopSynth`)
 * New NDFS based state space construction (`-mode ndfs`) with accepting cycle detection
 * New check-syntax mode (`-mode checksyntax`), that simply checks the syntax and terminates without doing any analysis.